Moved everything to ProjectController
